% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{applyIdentifiabilityConstraints}
\alias{applyIdentifiabilityConstraints}
\title{Apply identifiability constraints (sequential reallocation)}
\usage{
applyIdentifiabilityConstraints(coef, modelFit)
}
\arguments{
\item{coef}{Output of \code{\link{getCoefficients}}.}

\item{modelFit}{The object returned by \code{\link{trendModelFit}}.}
}
\value{
A list mirroring \code{coef} but adjusted.
}
\description{
Apply identifiability constraints (sequential reallocation)
}
