% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{applyIdentifiabilityConstraints2}
\alias{applyIdentifiabilityConstraints2}
\title{Apply identifiability constraints (one-shot projection)}
\usage{
applyIdentifiabilityConstraints2(coef, modelFit)
}
\arguments{
\item{coef}{Output of \code{\link{getCoefficients}}.}

\item{modelFit}{The object returned by \code{\link{trendModelFit}}.}
}
\value{
A list of adjusted components \code{ax}, \code{kt1}, \code{dx}, \code{dx_list}, \code{gc}.
}
\description{
Apply identifiability constraints (one-shot projection)
}
