% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{applyIdentifiabilityConstraintsTrends}
\alias{applyIdentifiabilityConstraintsTrends}
\title{Apply identifiability constraints (remove trends up - to a given year)}
\usage{
applyIdentifiabilityConstraintsTrends(coef, modelFit, lastYear = NULL)
}
\arguments{
\item{coef}{Output of \code{\link{getCoefficients}}.}

\item{modelFit}{The object returned by \code{\link{trendModelFit}}.}

\item{lastYear}{The last year up to which to remove trends. If NULL, use all years in \code{modelFit$years.fit}.}
}
\value{
A list of adjusted components \code{ax}, \code{kt1}, \code{dx}, \code{dx_list}, \code{gc}.
}
\description{
Apply identifiability constraints (remove trends up - to a given year)
}
