% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedRates.R
\name{fittedRatesTrendMoMo}
\alias{fittedRatesTrendMoMo}
\title{Function to extract fitted rates from a fitted trend MoMo model}
\usage{
fittedRatesTrendMoMo(fittedModel)
}
\arguments{
\item{fittedModel}{A fitted trend MoMo model object.}
}
\value{
A matrix of fitted rates with ages in rows and years in columns.
}
\description{
This function extracts the fitted rates from a fitted trend MoMo model object.
}
\details{
The general formulation implemented in this model is:
\deqn{
\log m_{x,t} = \alpha_x
+ \kappa^{(1)}_t
+ \sum_{j=1}^M \beta^{(j)}_x \kappa^{(j)}_t
+ d_x (t - \bar{t})
+ \sum_{i=1}^B d^{(i)}_x (t - t_i)_+
+ \gamma_{t-x},
}
}
