% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Extract structured coefficients from a fitted Trend-MoMo}
\usage{
getCoefficients(modelFit)
}
\arguments{
\item{modelFit}{The object returned by \code{\link{trendModelFit}}.}
}
\value{
A list with elements \code{ax}, \code{kt1}, \code{dx}, \code{dx_list}, \code{gc}.
}
\description{
Extract structured coefficients from a fitted Trend-MoMo
}
