% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTrendMoMo}
\alias{plotTrendMoMo}
\title{Plot adjusted Trend-MoMo coefficients}
\usage{
plotTrendMoMo(fittedModel)
}
\arguments{
\item{fittedModel}{The object returned by \code{\link{fitTrendMoMo}} or \code{\link{fitTrendMoMo2}}.}
}
\description{
Plot adjusted Trend-MoMo coefficients
}
