% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedRates.R
\name{predictLinkTrendMoMo}
\alias{predictLinkTrendMoMo}
\title{Function to predict log rates given a_x, k_t, d_x, d_xi, gc}
\usage{
predictLinkTrendMoMo(
  ax,
  kt1,
  dx,
  dx_list,
  gc,
  ages,
  years,
  trendChanges,
  years.fit = NULL
)
}
\arguments{
\item{ax}{Numeric vector of age-specific intercepts.}

\item{kt1}{Numeric vector of period factor.}

\item{dx}{Numeric vector of baseline trend components.}

\item{dx_list}{List of numeric vectors of trend break components.}

\item{gc}{Numeric vector of cohort effects.}

\item{ages}{Numeric vector of ages.}

\item{years}{Numeric vector of years.}

\item{trendChanges}{Numeric vector of calendar years at which trend breaks occur.}

\item{years.fit}{Optional numeric vector of years to calcuate tbar (if NULL, uses years).}
}
\value{
A matrix of predicted rates with ages in rows and years in columns.
}
\description{
This function predicts rates given the components of the trend MoMo model.
}
\details{
The general formulation implemented in this model is:
\deqn{
\log m_{x,t} = \alpha_x + \kappa^{(1)}_t +
\sum_{j=1}^M \beta^{(j)}_x \kappa^{(j)}_t +
d_x (t - \bar{t}) + \sum_{i=1}^B d^{(i)}_x (t - t_i)_+ + \gamma_{t-x}}
}
