% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateFutureRates}
\alias{simulateFutureRates}
\title{Simulate future mortality rates}
\usage{
simulateFutureRates(fittedModel, h = 20, nSim = 1, ...)
}
\arguments{
\item{fittedModel}{The object returned by \code{\link{fitTrendMoMo}} or \code{\link{fitTrendMoMo2}}.}

\item{h}{Numeric; the number of years ahead to simulate.}

\item{nSim}{Numeric; the number of simulation paths to generate.}

\item{...}{Additional parameters to pass to \code{\link{simulateKt1}} and \code{\link{simulateTrendDeviations}}.}
}
\value{
A list where each element corresponds to a simulation of future mortality rates.
Each element is a matrix of dimension \code{length(ages) x h} with ages in rows and years in columns.
}
\description{
This function simulates future mortality rates based on a fitted Trend-MoMo model.
}
