% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateKt1}
\alias{simulateKt1}
\title{Simulate kt1 future paths}
\usage{
simulateKt1(
  fittedModel,
  h = 20,
  nSim = 1,
  method_kt = c("normal", "rw"),
  method_params_kt = list(),
  ...
)
}
\arguments{
\item{fittedModel}{The object returned by \code{\link{fitTrendMoMo}} or \code{\link{fitTrendMoMo2}}.}

\item{h}{Numeric; the number of years ahead to simulate.}

\item{nSim}{Numeric; the number of simulation paths to generate.}

\item{method_kt}{Character; the simulation method to use. Options are:
\code{"rw"} for random walk, \code{"normal"} for normal distribution, more to be implemented.}

\item{method_params_kt}{A list of additional parameters for the simulation method.
For \code{method_kt = "rw"}, this can include:
\itemize{
\item \code{sd}: Numeric; the standard deviation of the random walk increments (default estimated from the data).
\item \code{last_kt1}: Numeric; the last observed value of kt1 to start the simulation from (default taken from the fitted model).
}
For \code{method_kt = "normal"}, this can include:
\itemize{
\item \code{mean}: Numeric; the mean of the normal distribution (default 0).
\item \code{sd}: Numeric; the standard deviation of the normal distribution (default estimated from the data).
}}
}
\value{
A matrix of dimension \code{h x nSim} where each column corresponds to a simulation of the kt1 path.
}
\description{
This function simulates future paths of the period effect kt1 for a Trend-MoMo model.
}
