% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateTrendDeviations}
\alias{simulateTrendDeviations}
\title{Function to simulat d^{(i)}_x terms}
\usage{
simulateTrendDeviations(
  fittedModel,
  h = 20,
  nSim = 1,
  method_size = c("normal", "AR1"),
  method_size_params = list(),
  method_duration = c("exponential", "gamma"),
  method_duration_params = list(),
  ...
)
}
\arguments{
\item{fittedModel}{The object returned by \code{\link{fitTrendMoMo}} or \code{\link{fitTrendMoMo2}}.}

\item{h}{Numeric; the number of years ahead to simulate.}

\item{nSim}{Numeric; the number of simulation paths to generate.}

\item{method_size}{Character; the simulation method to use. Options are:
\code{"normal"} for normal distribution, \code{"AR1"} for autoregressive model, possibly
with box cox transformation, more to be implemented.}

\item{method_size_params}{A list of additional parameters for the size simulation method.
For \code{method_size = "normal"}, this can include:
\itemize{
\item \code{mean}: Numeric; the mean of the normal distribution (default 0).
\item \code{sd}: Numeric; the standard deviation of the normal distribution (default 0.01).
}
For \code{method_size = "AR1"}, this can include:
\itemize{
\item \code{phi}: Numeric; the AR(1) coefficient (default 0.5).
\item \code{sd}: Numeric; the standard deviation of the white noise (default 0.01).
\item \code{init}: Numeric; the initial value of the AR(1) process (default 0).
\item \code{lambda} : Numeric; the box cox transformation parameter (default NULL, no transformation).
}}

\item{method_duration}{Character; the simulation method for duration. Options are:
\code{"exponential"} for exponential distribution, \code{"gamma"} for gamma distribution,
more to be implemented.}

\item{method_duration_params}{A list of additional parameters for the duration simulation method.
For \code{method_duration = "exponential"}, this can include:
\itemize{
\item \code{mean}: Numeric; the mean number of year between trend changes (default 10).
}
For \code{method_duration = "gamma"}, this can include:
\itemize{
\item \code{shape}: Numeric; the shape parameter of the gamma distribution (default 2).
\item \code{rate}: Numeric; the rate parameter of the gamma distribution (default 0.2).
}}
}
\value{
A list where each element corresponds to a simulation of the d^{(i)}_x terms. It has two
#' components:
\item{dx_list}{A list of length equal to the number of breakpoints in the model. Each element is a numeric
vector representing the simulated d^{(i)}_x values for that breakpoint.}
\item{breakpoints}{A numeric vector of the breakpoint years in the model.}
}
\description{
This funcion simulates future values of the age-specific trend deviation terms
d^{(i)}_x for a Trend-MoMo model with breakpoints
}
\details{
The d^{(i)}_x terms represent age-specific deviations from the linear trend
after each breakpoint t_i. This function simulates future values of these terms
based on a given model specification. In the simulation it assumes that the
future d^{(i)}_x terms do not depend on age, i.e., they are constant across ages.
}
